+++ packages/dso-toolkit/src/components/badge/badge.models.ts
@@ -1,6 +1,11 @@
 export type BadgeStatus = "primary" | "success" | "info" | "warning" | "error" | "outline" | "attention";
 
-export interface Badge {
+export type BadgeTooltipPlacement = "top" | "bottom" | "left" | "right";
+
+export interface Badge<TemplateFnReturnType> {
   status?: BadgeStatus;
   message: string;
+  label?: string;
+  toggletipPlacement?: BadgeTooltipPlacement;
+  children?: TemplateFnReturnType;
 }