+++ packages/react/src/components/badge/badge.react-template.tsx
@@ -1,14 +1,18 @@
 import { Badge } from "dso-toolkit";
-import * as React from "react";
+import React, { JSX } from "react";
 
 import { DsoBadge } from "../../components";
 import { ComponentImplementation } from "../../templates";
 
-export const reactBadge: ComponentImplementation<Badge> = {
+export const reactBadge: ComponentImplementation<Badge<JSX.Element>> = {
   component: "badge",
   implementation: "react",
   template: () =>
-    function badgeTemplate({ status, message }) {
-      return <DsoBadge status={status}>{message}</DsoBadge>;
+    function badgeTemplate({ status, message, label, toggletipPlacement, children }) {
+      return (
+        <DsoBadge status={status} label={label} toggletipPlacement={toggletipPlacement}>
+          {message} {children && <div slot="toggletip">{children}</div>}
+        </DsoBadge>
+      );
     },
 };