+++ storybook/src/components/badge/badge.core-template.ts
@@ -1,14 +1,20 @@
 import { Badge } from "dso-toolkit";
-import { html } from "lit-html";
+import { TemplateResult, html, nothing } from "lit-html";
 import { ifDefined } from "lit-html/directives/if-defined.js";
 
 import { ComponentImplementation } from "../../templates";
 
-export const coreBadge: ComponentImplementation<Badge> = {
+export const coreBadge: ComponentImplementation<Badge<TemplateResult>> = {
   component: "badge",
   implementation: "core",
   template: () =>
-    function badgeTemplate({ status, message }) {
-      return html`<dso-badge status=${ifDefined(status)}>${message}</dso-badge>`;
+    function badgeTemplate({ status, message, label, toggletipPlacement, children }) {
+      return html`<dso-badge
+        .status=${ifDefined(status)}
+        .label=${ifDefined(label)}
+        .toggletipPlacement=${ifDefined(toggletipPlacement)}
+      >
+        ${message} ${children ? html`<div slot="toggletip">${children}</div>` : nothing}
+      </dso-badge>`;
     },
 };